requirements.txt
# LangChain Core
langchain
langchain-core
langchain-groq


# Environment Variable Management
python-dotenv

# pip install -r requirements.txt



# extra
matplotlib
